Change EMUFILE fseek,ftell, and other related functions to use types that follow the C standard (long int for position offsets and size_t for sizes). This allows for larger max file sizes on unix platforms and cleaner code that better maps to the C standard.

+++ src/emufile.cpp
@@ -78,7 +78,7 @@ void EMUFILE_FILE::open(const char* fname, const char* mode)
 }
 
 
-void EMUFILE_FILE::truncate(s32 length)
+void EMUFILE_FILE::truncate(size_t length)
 {
 	::fflush(fp);
 	#ifdef _MSC_VER
